requirements.txt
Package                   Version
------------------------- ------------
altgraph                  0.17
cachetools                4.1.1
certifi                   2020.6.20
chardet                   3.0.4
click                     7.1.2
cssselect                 1.1.0
cssutils                  1.0.2
email-to                  0.1.0
emails                    0.6
future                    0.18.2
idna                      2.10
importlib-metadata        2.0.0
lxml                      4.5.2
Markdown                  3.3.3
mysql-connector-python    8.0.21
pefile                    2019.4.18
Pillow                    7.2.0
pip                       20.2.4
premailer                 3.7.0
protobuf                  3.13.0
pyinstaller               4.0
pyinstaller-hooks-contrib 2020.9
PyQt5                     5.15.1
PyQt5-sip                 12.8.1
pyqt5-tools               5.15.1.1.7.4
python-dateutil           2.8.1
python-dotenv             0.14.0
pywin32-ctypes            0.2.0
requests                  2.24.0
secure-smtplib            0.1.1
setuptools                47.1.0
six                       1.15.0
urllib3                   1.25.10
zipp                      3.4.0
